Streams should not be out of lambda handlers
