+++ lib/etl/handleError.js
@@ -25,7 +25,11 @@ const handleError = async (
     to: ['logs'],
   });
 
-  return callback(error);
+  if (callback) {
+    return callback(error);
+  }
+
+  throw error;
 };
 
 export default handleError;
