--- services/ingestion/etl/agri/csv/src/lib/parser.js
@@ -1,5 +1,0 @@
-import parse from 'csv-parse';
-
-const parser = parse({ columns: true });
-
-export default parser;
